refactor(graphql)!: `Manipulator::getPlaceholderTypeDefinitionNode()` => `AstManipulator::getOriginTypeDefinition()`.
